requirements.txt.txt
streamlit
pandas
scikit-learn
shap
matplotlib
scipy
fpdf2
